Drop redundant `void` `Return`
Most other nodes use an empty `void`-typed value to signify `void` return; we can do the same for returns by dropping `Return`. Also rename `ValueReturn` to just `Return`.
